+++ parlai/tasks/jsonfile/README.md
@@ -1,0 +1,7 @@
+# JsonFile Task Agent
+
+If you have a dialogue, QA or other text-only dataset that you can put in a text file in the format we will now describe, you can just load it directly from there, with no extra code!
+
+E.g. use with: 
+
+python parlai/scripts/display_data.py -t fromfile:parlaiformat --fromfile_datapath /tmp/data.txt
